+++ examples/next/faust-nx/package.json
@@ -9,15 +9,15 @@
   },
   "dependencies": {
     "@apollo/client": "^3.6.6",
+    "@wordpress/base-styles": "^4.7.0",
+    "@wordpress/block-library": "^7.13.0",
     "classnames": "^2.3.1",
     "faust-nx": "file:../../../packages/faust-nx/dist/index.js",
-    "faust-nx-cli": "file:../../../packages/faust-nx-cli/dist/index.js",
     "next": "^12.1.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "react": "^17.0.2"
   },
   "devDependencies": {
-    "@wordpress/base-styles": "^4.7.0",
-    "@wordpress/block-library": "^7.13.0"
+    "faust-nx-cli": "file:../../../packages/faust-nx-cli/dist/index.js"
   }
 }
