--- examples/next/faust-nx/pages/index.tsx
@@ -1,10 +1,0 @@
-import { getWordPressProps, WordPressTemplate } from 'faust-nx';
-import { GetStaticPropsContext } from 'next';
-
-export default function Page(props: any) {
-  return <WordPressTemplate {...props} />;
-}
-
-export function getStaticProps(ctx: GetStaticPropsContext) {
-  return getWordPressProps({ ctx });
-}
