+++ examples/next/faust-nx/pages/preview.js
@@ -1,5 +1,5 @@
 import { WordPressTemplate } from 'faust-nx';
 
-export default function Preview(props: any) {
+export default function Preview(props) {
   return <WordPressTemplate {...props} />;
 }
