--- examples/next/faust-nx/wp-templates/front-page.tsx
@@ -1,3 +1,0 @@
-export default function Component() {
-  return <>My home page</>;
-}
