+++ examples/next/faust-nx/wp-templates/index.js
@@ -1,9 +1,11 @@
-import page from './page';
 import category from './category';
+import frontPage from './front-page';
+import page from './page';
 import single from './single';
 
 export default {
-  page,
   category,
-  single
+  'front-page': frontPage,
+  page,
+  single,
 };
