--- examples/next/faust-nx/wp-templates/index.ts
@@ -1,11 +1,0 @@
-import page from './page';
-import archive from './archive';
-import single from './single';
-import frontPage from './front-page';
-
-export default {
-  page,
-  archive,
-  single,
-  'front-page': frontPage,
-};
